Start of the trainer, basic GUI work
This commit is the start of the trainer! I am going to focus working on this project and put my login bank on hold as I have been really interested in working with Java and a GUI. I do not have much experience with GUI's so I am learning swing in java through this project. This first commit is the basic skeleton of the main window. Next commit will most likely be the finished GUI.